+++ apps/desktop/src/lib/cmds/ext.ts
@@ -2,6 +2,8 @@ import { i18n } from "@/i18n"
 import { appState } from "@/stores"
 import { winExtMap } from "@/stores/winExtMap"
 import { helperAPI } from "@/utils/helper"
+import { paste } from "@/utils/hotkey"
+import { sleep } from "@/utils/time"
 import { trimSlash } from "@/utils/url"
 import { constructExtensionSupportDir } from "@kksh/api"
 import { db, spawnExtensionFileServer } from "@kksh/api/commands"
@@ -12,6 +14,7 @@ import { launchNewExtWindow, loadExtensionManifestFromDisk } from "@kksh/extensi
 import type { IKunkunFullServerAPI } from "@kunkunapi/src/api/server"
 import { convertFileSrc } from "@tauri-apps/api/core"
 import * as path from "@tauri-apps/api/path"
+import { getCurrentWindow } from "@tauri-apps/api/window"
 import * as fs from "@tauri-apps/plugin-fs"
 import { platform } from "@tauri-apps/plugin-os"
 import { goto } from "$app/navigation"
@@ -20,10 +23,12 @@ import * as v from "valibot"
 
 export const KunkunIframeExtParams = v.object({
 	url: v.string(),
+	cmdName: v.optional(v.string()),
 	extPath: v.string()
 })
 export type KunkunIframeExtParams = v.InferOutput<typeof KunkunIframeExtParams>
 export const KunkunTemplateExtParams = v.object({
+	url: v.optional(v.string()),
 	extPath: v.string(),
 	cmdName: v.string()
 })
@@ -36,10 +41,10 @@ export async function createExtSupportDir(extPath: string) {
 	}
 }
 
-function setTemplateExtParams(extPath: string, cmdName: string) {
+function setTemplateExtParams(extPath: string, cmdName: string, url?: string) {
 	localStorage.setItem(
 		"kunkun-template-ext-params",
-		JSON.stringify({ extPath, cmdName } satisfies KunkunTemplateExtParams)
+		JSON.stringify({ extPath, cmdName, url } satisfies KunkunTemplateExtParams)
 	)
 }
 
@@ -50,13 +55,15 @@ export async function onTemplateUiCmdSelect(
 ) {
 	await createExtSupportDir(ext.extPath)
 	const url = `/app/extension/ui-worker?extPath=${encodeURIComponent(ext.extPath)}&cmdName=${encodeURIComponent(cmd.name)}`
-	setTemplateExtParams(ext.extPath, cmd.name)
+	setTemplateExtParams(ext.extPath, cmd.name, url)
 	if (cmd.window) {
 		const winLabel = await winExtMap.registerExtensionWithWindow({ extPath: ext.extPath })
-		localStorage.setItem(
-			"kunkun-template-ext-params",
-			JSON.stringify({ url, extPath: ext.extPath } satisfies KunkunIframeExtParams)
-		)
+		const paramsStr = JSON.stringify({
+			url,
+			extPath: ext.extPath,
+			cmdName: cmd.name
+		} satisfies KunkunIframeExtParams)
+		localStorage.setItem("kunkun-template-ext-params", paramsStr)
 		const window = launchNewExtWindow(winLabel, url, cmd.window)
 		window.onCloseRequested(async (event) => {
 			await winExtMap.unregisterExtensionFromWindow(winLabel)
@@ -89,7 +96,21 @@ export async function onHeadlessCmdSelect(
 	}
 	const serverAPI: IKunkunFullServerAPI = constructJarvisServerAPIWithPermissions(
 		loadedExt.kunkun.permissions,
-		loadedExt.extPath
+		loadedExt.extPath,
+		{
+			recordSpawnedProcess: async (pid: number) => {
+				console.log("recordSpawnedProcess pid", pid)
+			},
+			getSpawnedProcesses: async () => {
+				console.log("getSpawnedProcesses")
+				return []
+			},
+			paste: async () => {
+				await getCurrentWindow().hide()
+				await sleep(200)
+				return paste()
+			}
+		}
 	)
 	const serverAPI2 = {
 		...serverAPI,