+++ packages/api/src/permissions/description.ts
@@ -17,6 +17,7 @@ export const permissionDescriptions: PermissionDescriptions = {
 	"clipboard:write-image": "Allows writing images to the clipboard",
 	"clipboard:read-files": "Allows reading copied files from the clipboard",
 	"clipboard:write-files": "Allows writing files to the clipboard",
+	"clipboard:paste": "Allows simulating paste operation",
 	"dialog:all": "Allows access to system dialog APIs, e.g. confirm, save, open, etc.",
 	"notification:all": "Allows sending system notifications",
 	"os:all": "Allows access to all operating system information",
@@ -24,6 +25,7 @@ export const permissionDescriptions: PermissionDescriptions = {
 	"shell:execute": "Allows executing shell commands",
 	"shell:spawn": "Allow spawning a new process and listen to the streaming of its output",
 	"shell:kill": "Allows killing processes by pid. Need this to kill the process you started.",
+	"shell:kill-any": "Allows killing any process by pid",
 	"shell:all":
 		"Grant all shell related permissions. Path scope and args regex validation is still required.",
 	"shell:stdin-write": "Allows writing to a command created.",