+++ packages/api/src/ui/template/index.ts
@@ -5,7 +5,6 @@
 // import { RPCChannel, WorkerChildIO, type DestroyableIoInterface } from "kkrpc/browser"
 import { RPCChannel, WorkerChildIO, type DestroyableIoInterface } from "kkrpc/browser"
 import type {
-	IClipboard,
 	IDialog,
 	// IEventInternal,
 	IFetchInternal,
@@ -27,6 +26,7 @@ import {
 } from "tauri-api-adapter/client"
 import type {
 	IApp,
+	IClipboard,
 	IDb,
 	IEvent,
 	IFs,