+++ apps/cli/CHANGELOG.md
@@ -1,5 +1,19 @@
 # kksh
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @kksh/api@0.1.5
+
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @kksh/api@0.1.4
+
 ## 0.1.1
 
 ### Patch Changes