+++ apps/create-kunkun/CHANGELOG.md
@@ -1,5 +1,19 @@
 # create-kunkun
 
+## 0.1.49
+
+### Patch Changes
+
+- Updated dependencies
+  - @kksh/api@0.1.5
+
+## 0.1.48
+
+### Patch Changes
+
+- Updated dependencies
+  - @kksh/api@0.1.4
+
 ## 0.1.45
 
 ### Patch Changes