+++ packages/api/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @kksh/api
 
+## 0.1.5
+
+### Patch Changes
+
+- Add clipboard.paste() API
+
+## 0.1.4
+
+### Patch Changes
+
+- Add killPid extension API
+
 ## 0.1.2
 
 ### Patch Changes