+++ packages/extensions/demo-worker-template-ext/CHANGELOG.md
@@ -1,5 +1,19 @@
 # demo-template-extension
 
+## 0.0.11
+
+### Patch Changes
+
+- Updated dependencies
+  - @kksh/api@0.1.5
+
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @kksh/api@0.1.4
+
 ## 0.0.9
 
 ### Patch Changes