+++ packages/templates/template-ext-react/CHANGELOG.md
@@ -1,5 +1,19 @@
 # template-ext-react
 
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @kksh/api@0.1.5
+
+## 0.0.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @kksh/api@0.1.4
+
 ## 0.0.8
 
 ### Patch Changes